requirements.txt
typer>=0.9.0
anthropic>=0.31.0
claude-agent-sdk>=0.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0
mcp>=1.0.0

# Test dependencies
pytest>=9.0.0
pytest-asyncio>=1.0.0
pytest-mock>=3.0.0
respx>=0.22.0
